+++ backend/src/main/java/mzumot/plantsapp/backend/repository/PlantsRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import mzumot.plantsapp.backend.model.Plant;
 
 @Repository
-public interface PlantsRepository extends JpaRepository<Plant, Integer> {
+public interface PlantsRepository extends JpaRepository<Plant, Long> {
  
     List<Plant> findAll();
